requirements.txt
numpy
opencv-python
scipy
photutils
Pillow
pikepdf
argparse
matplotlib
tqdm